Time: 27 ms (5.25%) | Memory: 44.6 MB (7.30%) - LeetSync

+++ 9-palindrome-number/palindrome-number.java
@@ -2,7 +2,7 @@ class Solution {
     public boolean isPalindrome(int x) {
         if(x<0) return false;
         int ip = x;
-        long rev=0;
+        int rev=0;
         while(x>0){
             rev=rev*10+(x%10);
             x=x/10;
